requirements.txt
openai
PyYAML
argcomplete
pytest
